add acceleration + move level.c defines to defines.h

+++ lib/level.h
@@ -6,5 +6,5 @@
 
 void draw_level_to_screen(char (*foreground)[GRID_Y]);
 char** generate_screen(int stage);
-void scroll_level(int distance, char (*foreground)[GRID_Y]);
+void scroll_level(float distance, char (*foreground)[GRID_Y]);
 bool should_generate_screen(char (*level)[GRID_Y]);
